adding sample 0 and 1
