+++ samples/s00-rayGenOnly/CMakeLists.txt
@@ -22,16 +22,16 @@
 
 embed_devicecode(
   OUTPUT_TARGET
-    int05_deviceCode
+    s00_deviceCode
   SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/deviceCode.hlsl
+    # For whatever reason, including a .h in the spv causes a VK_INITIALIZATION_FAILED (-3) when creating shader modules
+    # ${CMAKE_CURRENT_SOURCE_DIR}/deviceCode.h
 )
 
-add_executable(int05_computePrimitive hostCode.cpp)
-target_link_libraries(int05_computePrimitive
+add_executable(s00_rayGenOnly hostCode.cpp)
+target_link_libraries(s00_rayGenOnly
   PRIVATE
-    int05_deviceCode
+    s00_deviceCode
     gprt::gprt
-    glfw
-    ${OPENGL_gl_LIBRARY}
 )
