+++ samples/s00-rayGenOnly/deviceCode.h
