+++ samples/s00-rayGenOnly/hostCode.cpp
@@ -35,10 +35,10 @@
   std::cout << "#gprt.sample(main): " << message << std::endl;   \
   std::cout << GPRT_TERMINAL_DEFAULT;
 
-extern GPRTProgram int00_deviceCode;
+extern GPRTProgram s00_deviceCode;
 
 // initial image resolution
-const int2 fbSize = {800,600};
+const int2 fbSize = {1400, 460};
 
 // final image output
 const char *outFileName = "s00-rayGenOnly.png";
@@ -47,28 +47,34 @@ const char *outFileName = "s00-rayGenOnly.png";
 int main(int ac, char **av)
 {
   // The output window will show comments for many of the methods called.
-  // Walking through the code line by line with a debugger is educational.
+  // We recommend walking through the code line by line with a debugger. 
   LOG("gprt example '" << av[0] << "' starting up");
 
   // ##################################################################
-  // set up all the *CODE* we want to run
+  // set up all the GPU kernels we want to run
   // ##################################################################
 
   LOG("building module, programs, and pipeline");
 
-  // Since we'll be using a window, we request support for an image swapchain
-  gprtRequestWindow(fbSize.x, fbSize.y, "Int00 Raygen Only");
+  // Since we'll be using a window, we request support for one here. 
+  // If a window can't be made, we can still use GPRT, but a window wont appear. 
+  // And if you know you don't need a window, then this call is optional.
+  gprtRequestWindow(fbSize.x, fbSize.y, "S00 Raygen Only");
 
-  // Initialize Vulkan, and create a "gprt device," a context to hold the
-  // ray generation shader and output buffer. The "1" is the number of devices requested.
-  GPRTContext gprt = gprtContextCreate(nullptr, 1);
+  // Initialize GPRT. Under the hood, this identifies all compatible GPUs (ones
+  // that support ray tracing) selects from that compatible list. If a window
+  // was requested, one will appear after this call. 
+  GPRTContext context = gprtContextCreate();
 
-  // SPIR-V is the intermediate code that the GPU deviceCode.hlsl shader program is converted into.
+  // Device code is represented through SPIR-V, which is the intermediate code 
+  // that the GPU deviceCode.hlsl shader program is compiled into.
   // You can see the machine-centric SPIR-V code in
-  // build\samples\cmd00-rayGenOnly\deviceCode.spv
+  // build\s00-rayGenOnly\deviceCode.spv
   // We store this SPIR-V intermediate code representation in a GPRT module.
-  GPRTModule module = gprtModuleCreate(gprt,int00_deviceCode);
+  GPRTModule module = gprtModuleCreate(context, s00_deviceCode);
 
+  // All ray tracing programs start off with a "Ray Generation" kernel.
+  // All "parameters" we'll pass to that ray generation kernel are defined here.
   GPRTVarDecl rayGenVars[]
     = {
       { "fbPtr", GPRT_BUFFER, GPRT_OFFSETOF(RayGenData, fbPtr) },
@@ -80,45 +86,67 @@ int main(int ac, char **av)
   // Allocate room for one RayGen shader, create it, and
   // hold on to it with the "gprt" context
   GPRTRayGen rayGen
-      = gprtRayGenCreate(gprt, module, "simpleRayGen",
+      = gprtRayGenCreate(context, module, "simpleRayGen",
                       sizeof(RayGenData),rayGenVars,-1);
 
-  // (re-)builds all vulkan programs, with current pipeline settings
-  gprtBuildPipeline(gprt);
+  // (re-)builds all gprt programs, with current pipeline settings
+  gprtBuildPipeline(context);
 
-  // ------------------------------------------------------------------
-  // alloc buffers
-  // ------------------------------------------------------------------
+  // ##################################################################
+  // allocating GPU buffers
+  // ##################################################################
+  
+  // When working with GPUs, we work with two different pools of memory:
+  // traditional CPU memory and GPU memory. GPU programs can only read
+  // from GPU buffers of memory. The CPU can temporarily read and write
+  // to GPU bufer memory when that buffer is "mapped" to the CPU.
+
+  // Our frame buffer here will be used to hold pixel color values
+  // that we'll present to the window and save to an image
   LOG("allocating frame buffer");
-  GPRTBuffer frameBuffer = gprtDeviceBufferCreate(gprt,
-                                          /*type:*/GPRT_INT,
-                                          /*size:*/fbSize.x*fbSize.y);
+  GPRTBuffer frameBuffer = gprtDeviceBufferCreate(
+                              context,
+                              GPRT_INT, /* type */ 
+                              fbSize.x*fbSize.y /* size in count */);
+  // See also gprtHostBufferCreate and gprtSharedBufferCreate.
+  // To get a pointer to this buffer, call gprtBufferMap, then 
+  // gprtBufferGetPointer. When done writing to the pointer, call 
+  // gprtBufferUnmap.
 
-  // ------------------------------------------------------------------
-  // build Shader Binding Table (SBT) required to trace the groups
-  // ------------------------------------------------------------------
+  // ##################################################################
+  // Build the shader binding table
+  // ##################################################################
+  
+  // The shader binding table is used by the GPU to map parameters to GPU 
+  // kernels. 
+
   gprtRayGenSet3f(rayGen,"color0",0.1f,0.1f,0.1f);
   gprtRayGenSet3f(rayGen,"color1",.0f,.0f,.0f);
   gprtRayGenSetBuffer(rayGen,"fbPtr",frameBuffer);
   gprtRayGenSet2i(rayGen,"fbSize",fbSize.x,fbSize.y);
+  
   // Build a shader binding table entry for the ray generation record.
-  gprtBuildShaderBindingTable(gprt);
+  gprtBuildShaderBindingTable(context, GPRT_SBT_RAYGEN);
 
 
   // ##################################################################
-  // now that everything is ready: launch it ....
+  // Launching
   // ##################################################################
+
+  // To execute code on the GPU, we use "gprtRayGenLaunch". Only ray generation
+  // and compute programs can be launched. 
+  // ( compute programs will be covered later :) )
   LOG("executing the launch ...");
   do 
   {
     // Calls the GPU raygen kernel function
-    gprtRayGenLaunch2D(gprt,rayGen,fbSize.x,fbSize.y);
+    gprtRayGenLaunch2D(context,rayGen,fbSize.x,fbSize.y);
     
-    // If a window exists, presents the framebuffer here to that window
-    gprtBufferPresent(gprt, frameBuffer); 
+    // If a window exists, presents the frame buffer here to that window
+    gprtBufferPresent(context, frameBuffer); 
   } 
   // returns true if "X" pressed or if in "headless" mode
-  while (!gprtWindowShouldClose(gprt)); 
+  while (!gprtWindowShouldClose(context)); 
 
   // Save final frame to an image
   LOG("done with launch, writing frame buffer to " << outFileName);
@@ -133,7 +161,10 @@ int main(int ac, char **av)
   gprtBufferDestroy(frameBuffer);
   gprtRayGenDestroy(rayGen);
   gprtModuleDestroy(module);
-  gprtContextDestroy(gprt);
+
+  // If a window was made, this call is the one that will actually destroy 
+  // that window. 
+  gprtContextDestroy(context);
 
   LOG_OK("seems all went OK; app is done, this should be the last output ...");
 }
